refactor conans-conan userio
