+++ conans/client/rest/remote_credentials.py
@@ -4,11 +4,11 @@ import platform
 
 from jinja2 import Template
 
+from conan.api.input import UserInput
 from conan.internal.cache.home_paths import HomePaths
 
 from conans.client.loader import load_python_file
 from conan.api.output import ConanOutput
-from conans.client.userio import UserInput
 from conan.internal.errors import scoped_traceback
 from conan.errors import ConanException
 from conans.util.files import load
